% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_graph_list.R
\name{make_graph_list}
\alias{make_graph_list}
\title{A way to make a list for further use in other helper functions}
\usage{
make_graph_list(.data)
}
\arguments{
\item{.data}{a data frame made in another helper function}
}
\value{
a list of the variable list_names
}
\description{
A way to make a list for further use in other helper functions
}
