% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count_skeleton.R
\name{plot_count_skeleton}
\alias{plot_count_skeleton}
\title{plot_count_skeleton : mMake the skeleton for all plots for count metrics}
\usage{
plot_count_skeleton(.data, list_name)
}
\arguments{
\item{.data}{a data frame cleaned up from df_count_metric_values.rda}

\item{list_name}{a name found in df_count_metric_values.rda}
}
\value{
a plot
}
\description{
This function generate a plot with year on the x-axis and count values on the y-axis, including text labels
and a legend.
}
