% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rate_regression_skeleton.R
\name{plot_rate_regression_skeleton}
\alias{plot_rate_regression_skeleton}
\title{plot_rate_regression_skeleton() : skeleton function to create retention regression plot for
"utMetrics Dashboard" shiny app.}
\usage{
plot_rate_regression_skeleton(.data, list_name)
}
\arguments{
\item{.data}{a data frame cleaned up from df_rate_metric_values.rda}

\item{list_name}{a name list found in df_rate_metric_values.rda}
}
\value{
a plot
}
\description{
This function is a skeleton for creating a plot to visualize retention regression over time. It fits
a linear regression model to the provided data and plots the regression line along with the data
points. The plot also includes the predicted rate for the next year, and labels for the residuals
and prediction.
}
