% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rate_skeleton.R
\name{plot_rate_skeleton}
\alias{plot_rate_skeleton}
\title{make the skeleton for all plots for rate metrics}
\usage{
plot_rate_skeleton(.data, list_name)
}
\arguments{
\item{.data}{a data frame cleaned up from df_rate_metric_values.rda}

\item{list_name}{a name found in df_rate_metric_values.rda}
}
\value{
a plot
}
\description{
make the skeleton for all plots for rate metrics
}
