% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sum_regression_skeleton.R
\name{plot_sum_regression_skeleton}
\alias{plot_sum_regression_skeleton}
\title{plot_sum_regression_skeleton() : skeleton function to create retention regression plot for
"utMetrics Dashboard" shiny app}
\usage{
plot_sum_regression_skeleton(.data, list_name)
}
\arguments{
\item{.data}{a data frame cleaned up from df_sum_metric_values.rda}

\item{list_name}{a name list found in df_rate_metric_values.rda}
}
\value{
a plot
}
\description{
plot_sum_regression_skeleton() : skeleton function to create retention regression plot for
"utMetrics Dashboard" shiny app
}
