% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sum_skeleton.R
\name{plot_sum_skeleton}
\alias{plot_sum_skeleton}
\title{plot_sum_skeleton() : make the skeleton for all plots for sum metrics}
\usage{
plot_sum_skeleton(.data, list_name)
}
\arguments{
\item{.data}{a data frame cleaned up from df_sum_metric_values.rda}

\item{list_name}{a name found in df_sum_metric_values.rda}
}
\value{
a plot
}
\description{
This function generate a plot with year on the x-axis and sum values on the y-axis, including text labels and a legend.
}
