% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_count_interacitve_plots.R
\name{utm_count_interacitve_plots}
\alias{utm_count_interacitve_plots}
\title{utm_count_interacitve_plots() : Generate interactive count plots using the plotly and purr libraries}
\usage{
utm_count_interacitve_plots(.data)
}
\arguments{
\item{.data}{a data frame from df_count_metric_values.rda}
}
\value{
many plotly graphs of count metrics
}
\description{
This function creates multiple interactive plots for count metrics using utm_count_interactive_plot().
It processes data frames from df_count_metric_values.rda.
}
