% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_count_interactive_plot.R
\name{utm_count_interactive_plot}
\alias{utm_count_interactive_plot}
\title{utm_count_interactive_plot () : For one interactive plot at-a-time using plotly, for count metric}
\usage{
utm_count_interactive_plot(.data, list_name)
}
\arguments{
\item{.data}{A data frame from df_count_metric_values.rda}

\item{list_name}{a var in df_count_metric_values.rda}
}
\value{
An interactive plot for count
}
\description{
The function takes a data frame from df_count_metric_values.rda, creates a plot using plot_count_skeleton() function,
and increases interactivity with tooltips showing text and count values.
}
