% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_count_interactive_tables.R
\name{utm_count_interactive_tables}
\alias{utm_count_interactive_tables}
\title{utm_count_interactive_tables () : For many interactive table using the DT library and purrr for a count metrics.}
\usage{
utm_count_interactive_tables(.data)
}
\arguments{
\item{.data}{a data frame from df_count_metric_values.rda}
}
\value{
many DT tables of count metrics
}
\description{
This function converts title names to snake case, creates a list of interactive tables, and returns the tables by using
"utm_count_interactive_table()" function.
}
