% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_count_shiny_table.R
\name{utm_count_shiny_table}
\alias{utm_count_shiny_table}
\title{utm_count_shiny_table() : Create count table for "utMetrics Dashboard" shiny app.}
\usage{
utm_count_shiny_table(.data)
}
\arguments{
\item{.data}{A data frame from df_count_metric_values.rda}
}
\value{
A table for count. It consists of three columns: Year, Count and Category.
}
\description{
This function effectively shows the count table in the utMetrics dashboard. It takes filtered data frame
as input, creates title and used the function utm_count_interactive_table()' to produce the table. For one
table at-a-time using DT library.
}
