% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_count_static_plots.R
\name{utm_count_static_plots}
\alias{utm_count_static_plots}
\title{utm_count_static_plots() : For many static plots using the plotly and purr libraries for a count metrics.}
\usage{
utm_count_static_plots(.data)
}
\arguments{
\item{.data}{a data frame from df_count_metric_values.rda}
}
\value{
plots of count metrics
}
\description{
This function converts title names to snake case, creates a list of plots, and returns the plots using
utm_count_static_plot() function.
}
