% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_count_static_regression_plot.R
\name{utm_count_static_regression_plot}
\alias{utm_count_static_regression_plot}
\title{utm_count_static_regression_plot() : Modify headcount regression plot for "utMetrics Dashboard" shiny app.}
\usage{
utm_count_static_regression_plot(.data, list_name)
}
\arguments{
\item{.data}{a data frame cleaned up from df_count_metric_values.rda}

\item{list_name}{a name found in df_count_metric_values.rda}
}
\value{
A plot that uses plot_count_regression_skeleton as its base
}
\description{
This function utilizes the plot_count_regression_skeleton function to create the plot and then removes the legend
from the plot.
}
