% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_count_static_regression_plots.R
\name{utm_count_static_regression_plots}
\alias{utm_count_static_regression_plots}
\title{For many static plots using the ggplot2 and purr libraries for a count regression metrics.}
\usage{
utm_count_static_regression_plots(.data)
}
\arguments{
\item{.data}{a data frame from df_count_metric_values.rda}
}
\value{
many plots of count regression metrics
}
\description{
For many static plots using the ggplot2 and purr libraries for a count regression metrics.
}
