% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_count_static_table.R
\name{utm_count_static_table}
\alias{utm_count_static_table}
\title{For one static table at-a-time using the DT library.}
\usage{
utm_count_static_table(.data, list_name)
}
\arguments{
\item{.data}{a data frame made from df_count_metric_values.rda}

\item{list_name}{a distinct var in df_count_metric_values.rda}
}
\value{
A gt table
}
\description{
For one static table at-a-time using the DT library.
}
