% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_count_static_tables.R
\name{utm_count_static_tables}
\alias{utm_count_static_tables}
\title{For many static tables using the gt and purrr libraries for a count metrics.}
\usage{
utm_count_static_tables(.data)
}
\arguments{
\item{.data}{a data frame from df_count_metric_values.rda}
}
\value{
many gt tables of count metrics
}
\description{
For many static tables using the gt and purrr libraries for a count metrics.
}
