% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_export_as_plots.R
\name{utm_export_as_plots}
\alias{utm_export_as_plots}
\title{Produce a folder with .png files made from the plot functions}
\usage{
utm_export_as_plots(list_name)
}
\arguments{
\item{list_name}{the name made from either the plot_graph_for_rate or plot_graph_for_count functions}
}
\value{
A folder with .png files
}
\description{
Produce a folder with .png files made from the plot functions
}
