% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_export_tables.R
\name{utm_export_tables}
\alias{utm_export_tables}
\title{Produce a folder with ,png files made from the gt table functions}
\usage{
utm_export_tables(list_name)
}
\arguments{
\item{list_name}{the name made from either the ut_outcome_count_gt_tables or ut_outcome_rate_gt_tables functions}
}
\value{
a folder called tables filled with .png
}
\description{
Produce a folder with ,png files made from the gt table functions
}
