% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_rate_interactive_plots.R
\name{utm_rate_interacitve_plots}
\alias{utm_rate_interacitve_plots}
\title{utm_rate_interacitve_plots() : Generate interactive rate plots using the plotly and purr libraries}
\usage{
utm_rate_interacitve_plots(.data)
}
\arguments{
\item{.data}{a data frame from df_rate_metric_values.rda}
}
\value{
many plotly graphs of rate metrics
}
\description{
This function creates multiple interactive plots for rate metrics using utm_rate_interactive_plot().
It processes data frames from df_rate_metric_values.rda.
}
