% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_rate_interactive_regression_plot.R
\name{utm_rate_interactive_regression_plot}
\alias{utm_rate_interactive_regression_plot}
\title{For one interactive plot at-a-time using ggplot2 and plotly, for a rate regression metrics.}
\usage{
utm_rate_interactive_regression_plot(.data, list_name)
}
\arguments{
\item{.data}{a data frame cleaned up from df_rate_metric_values.rda}

\item{list_name}{a name found in df_rate_metric_values.rda}
}
\value{
A plot that uses plot_rate_regression_skeleton as its base
}
\description{
For one interactive plot at-a-time using ggplot2 and plotly, for a rate regression metrics.
}
