% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_rate_interactive_regression_plots.R
\name{utm_rate_interactive_regression_plots}
\alias{utm_rate_interactive_regression_plots}
\title{For many static plots using the plotly and purr libraries for a rate regression metrics.}
\usage{
utm_rate_interactive_regression_plots(.data)
}
\arguments{
\item{.data}{a data frame from df_rate_metric_values.rda}
}
\value{
many plots of rate regression metrics
}
\description{
For many static plots using the plotly and purr libraries for a rate regression metrics.
}
