% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_rate_static_regresssion_plot.R
\name{utm_rate_regression_plot}
\alias{utm_rate_regression_plot}
\title{utm_rate_regression_plot () : Creates a static Regression Plot}
\usage{
utm_rate_regression_plot(.data, list_name)
}
\arguments{
\item{.data}{a data frame cleaned up from df_rate_metric_values.rda}

\item{list_name}{a name found in df_rate_metric_values.rda}
}
\value{
A plot that uses plot_rate_regression_skeleton as its base
}
\description{
This function generates a regression plot by using  "plot_rate_regression_skeleton()" function filtered by a
specified list name, and removes the legend for a cleaner presentation.
}
