% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_rate_shiny_plot.R
\name{utm_rate_shiny_plot}
\alias{utm_rate_shiny_plot}
\title{utm_rate_shiny_plot() : For one interactive plot at-a-time using plotly, for shiny dashboard}
\usage{
utm_rate_shiny_plot(.data)
}
\arguments{
\item{.data}{A data frame from df_count_rate_values.rda}
}
\value{
An interactive plot for count
}
\description{
This function creates of one interactive plot at a time using Plotly for rate metric. It takes a data frame from
df_rate_metric_values.rda, processes it to generate an interactive plot, and returns the plot for shiny dashboard.
}
