% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_rate_shiny_regression.R
\name{utm_rate_shiny_regression}
\alias{utm_rate_shiny_regression}
\title{utm_rate_shiny_regression() : Modify retention regression plot for "utMetrics Dashboard" shiny app.}
\usage{
utm_rate_shiny_regression(.data)
}
\arguments{
\item{.data}{A data frame from df_rate_metric_values.rda}
}
\value{
An regression plot for rate
}
\description{
This function effectively shows the retention regression plot in the utMetrics dashboard. It takes filtered data frame
as input, creates title and used the function 'utm_rate_regression_plot()' to produce the plot. For one
plot at-a-time.
}
