% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_rate_shiny_table.R
\name{utm_rate_shiny_table}
\alias{utm_rate_shiny_table}
\title{utm_rate_shiny_table() : Create retention table for "utMetrics Dashboard" shiny app.}
\usage{
utm_rate_shiny_table(.data)
}
\arguments{
\item{.data}{A data frame from df_rate_metric_values.rda}
}
\value{
A table for rate. It consists of three columns: Year, Outcome Rate and Category.
}
\description{
This function effectively shows the retention table in the utMetrics dashboard. It takes filtered data frame
as input, creates title and used the function utm_rate_interactive_table()' to produce the table. For one
table at-a-time using DT library.
}
