% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_rate_static_plot.R
\name{utm_rate_static_plot}
\alias{utm_rate_static_plot}
\title{utm_rate_static_plot() : Generate static rate plot}
\usage{
utm_rate_static_plot(.data, list_name)
}
\arguments{
\item{.data}{a data frame cleaned up from df_rate_metric_values.rda}

\item{list_name}{a name found in df_rate_metric_values.rda}
}
\value{
A plot that uses plot_rate_skeleton as its base
}
\description{
This function produces a static plot for a specified rate metric based on the provided dataset and list name.
It filters the dataset to select data associated with the provided list name and extracts the title for the plot.
The plot is generated using "plot_rate_skeleton()" as its base, with legend customization based on the presence of
a colon in the title.
}
