% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_rate_static_regression_plots.R
\name{utm_rate_static_regression_plots}
\alias{utm_rate_static_regression_plots}
\title{utm_rate_static_regression_plots () : Creates static Regression Plots}
\usage{
utm_rate_static_regression_plots(.data)
}
\arguments{
\item{.data}{a data frame from df_rate_metric_values.rda}
}
\value{
many plots of rate regression metrics
}
\description{
This function generates multiple static regression plots by transforming the data and applying
the \code{utm_rate_static_regression_plot()} function to each subset of data, identified by unique list names.
}
