% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_sum_interacitve_plots.R
\name{utm_sum_interacitve_plots}
\alias{utm_sum_interacitve_plots}
\title{utm_sum_interacitve_plots() : Generate interactive sum plots using the plotly and purr libraries}
\usage{
utm_sum_interacitve_plots(.data)
}
\arguments{
\item{.data}{a data frame from df_sum_metric_values.rda}
}
\value{
many plotly graphs of sum metrics
}
\description{
This function creates multiple interactive plots for sum metrics using utm_sum_interactive_plot().
It processes data frames from df_sum_metric_values.rda.
}
