% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_sum_interactive_tables.R
\name{utm_sum_interactive_tables}
\alias{utm_sum_interactive_tables}
\title{utm_sum_interactive_tables () : For many interactive table using the DT library and purrr for a sum metrics.}
\usage{
utm_sum_interactive_tables(.data)
}
\arguments{
\item{.data}{a data frame from df_sum_metric_values.rda}
}
\value{
many DT tables of sum metrics
}
\description{
This function converts title names to snake case, creates a list of interactive tables, and returns the tables
named by their list names.
}
