% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_sum_shiny_regression.R
\name{utm_sum_shiny_regression}
\alias{utm_sum_shiny_regression}
\title{utm_sum_shiny_regression() : Modify sum regression plot for "utMetrics Dashboard" shiny app.}
\usage{
utm_sum_shiny_regression(.data)
}
\arguments{
\item{.data}{A data frame from df_sum_metric_values.rda}
}
\value{
An regression plot for sum
}
\description{
This function effectively shows the fte(sum) regression plot in the utMetrics dashboard. It takes filtered data frame
as input, creates title and used the function 'utm_sum_static_regression_plot()' to produce the plot. For one
plot at-a-time.
}
