% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_sum_shiny_table.R
\name{utm_sum_shiny_table}
\alias{utm_sum_shiny_table}
\title{utm_sum_shiny_table() : Create sum table for "utMetrics Dashboard" shiny app.}
\usage{
utm_sum_shiny_table(.data)
}
\arguments{
\item{.data}{A data frame from df_count_metric_values.rda}
}
\value{
A table for sum (fte). It consists of three columns: Year, sum and Category.
}
\description{
This function effectively shows the sum table in the utMetrics dashboard. It takes filtered data frame
as input, creates title and used the function utm_sum_interactive_table()' to produce the table.
}
