% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_sum_static_plot.R
\name{utm_sum_static_plot}
\alias{utm_sum_static_plot}
\title{utm_sum_static_plot() : One static plot at-a-time using ggplot2, for a sum metrics.}
\usage{
utm_sum_static_plot(.data, list_name)
}
\arguments{
\item{.data}{a data frame cleaned up from df_sum_metric_values.rda}

\item{list_name}{a name found in df_sum_metric_values.rda}
}
\value{
A plot that uses plot_sum_skeleton as its base
}
\description{
This function adjusts legend position based on the plot title format and prints the plot
}
