% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_sum_static_plots.R
\name{utm_sum_static_plots}
\alias{utm_sum_static_plots}
\title{utm_sum_static_plots() : For many static plots using the plotly and purr libraries for a sum metrics.}
\usage{
utm_sum_static_plots(.data)
}
\arguments{
\item{.data}{a data frame from df_sum_metric_values.rda}
}
\value{
plots of sum metrics
}
\description{
This function converts title names to snake case, creates a list of plots, and returns the plots using
utm_sum_static_plot() function.
}
