% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_sum_static_regression_plot.R
\name{utm_sum_static_regression_plot}
\alias{utm_sum_static_regression_plot}
\title{utm_sum_static_regression_plot() : Modify sum regression plot for "utMetrics Dashboard" shiny app.}
\usage{
utm_sum_static_regression_plot(.data, list_name)
}
\arguments{
\item{.data}{a data frame cleaned up from df_sum_metric_values.rda}

\item{list_name}{a name found in df_sum_metric_values.rda}
}
\value{
A plot that uses plot_sum_regression_skeleton as its base
}
\description{
This function utilizes the "plot_sum_regression_skeleton()" function to create the plot and then removes the legend
from the plot.
}
