% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_sum_static_table.R
\name{utm_sum_static_table}
\alias{utm_sum_static_table}
\title{utm_sum_static_table() : Create a Formatted GT Table}
\usage{
utm_sum_static_table(.data, list_name)
}
\arguments{
\item{.data}{a data frame made from df_sum_metric_values.rda}

\item{list_name}{a distinct var in df_sum_metric_values.rda}
}
\value{
A gt table
}
\description{
This function generates a formatted GT table from a given data frame, filtering by a specified list name and
renaming columns for better readability.
}
